添加新功能sys，数据处理部分修整完毕
